% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/song.datasets.R
\name{chickadees}
\alias{chickadees}
\title{A sample interaction from black-capped chickadees.}
\format{A data frame with 50 rows and 3 variables}
\source{
Baker TM, Wilson DR, Mennill DJ. 2013. Vocal signals predict attack
during aggressive interactions in black-capped chickadees. Anim Behav. 84:
965-974.
}
\description{
A data set containing the start and end times of each song produced during
an interaction between a black-capped chickadee and a fixed-interval
playback stimulus. These data are from a study by Baker, Wilson and Mennill
(2013) designed to evaluate whether song overlap predicts the attack of a
rival in this species.
}

