% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/song.R
\name{function_song}
\alias{function_song}
\title{A reserve song.}
\usage{
function_song(a = 99, liquid = "beer", vessel = "bottles",
  surface = "wall")
}
\arguments{
\item{a}{The number of the verses. Defaults to 99.}

\item{liquid}{The liquid. Defaults to beer.}

\item{vessel}{The vessel. Defaults to bottles.}

\item{surface}{The surface. Defaults to wall.}
}
\value{
A song of \code{a} \code{vessel} of \code{liquid} on the \code{surface}
}
\description{
This function prints a reverse song.
}
\examples{
function_song(3,'milk','boxes','wall')
function_song(10,'latte','cups','table')
}
