% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/song.R
\docType{data}
\name{manakins}
\alias{manakins}
\title{An interaction between neighboring pairs of long-tailed manakins.}
\format{A data frame with 50 rows and 3 variables}
\source{
Maynard DF, Ward KA, Doucet SM, Mennill DJ. 2012. Calling in an acoustically
competitive environment: duetting male long-tailed manakins avoid
overlapping neighbours but not playback-simulated rivals. Anim Behav.
84:563-573.
}
\usage{
manakins
}
\description{
A data set containing the start and end times of each song produced during
an interaction between neighboring male-male pairs of long-tailed manakins.
These data are from a study by Maynard, Ward, Doucet, and Mennill (2012)
designed to evaluate whether male-male pairs vary the timing of their duets
with respect to the duets of males in adjacent leks.
}
\keyword{datasets}

