% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/song.accuracy.R
\name{song.AccuracyTest}
\alias{song.AccuracyTest}
\title{Test the accuracy of a song overlap null model.}
\usage{
song.AccuracyTest(rndm.pairs, num.rand, overlap.function, randomize.function)
}
\arguments{
\item{rndm.pairs}{A list created by \code{\link{song.BatchReadSongs}} that
contains the performance statistics of non-interacting pairs.}

\item{num.rand}{A numeric value indicating the desired number of
randomizations.}

\item{overlap.function}{The function to be used to calculate the amount of
overlap. \code{\link{song.TimeOverlap}} returns the duration of overlap
in seconds; \code{\link{song.NumOverlap}} returns the number of overlapping
songs.}

\item{randomize.function}{The function to be used to generate the randomized
performances for each individual. Options include
\code{\link{song.RandomizeSampleGaps}},
\code{\link{song.RandomizeKeepGaps}}, and
\code{\link{song.RandomizeKeepSongOrder}}.}
}
\value{
\code{song.AccuracyTest} returns a list containing the results of the
simulations (see \code{\link{song.Simulate}} for a description of the
output) and a named vector containing the following components:
\describe{
  \item{Average Error}{The average difference between the actual and
  predicted amounts of overlap.}
  \item{Average Percent Error}{The average error expressed as a percent of
  the total interaction. If overlap was measured as a duration, this value
  represents the average error divided by the total duration of the
  interaction. If overlap was measured as the number of overlapping songs,
  this value represents the average error divided by the total number of
  songs in the interaction.}
  \item{P < 0.025}{The number of pairs for which overlapping occurred
  significantly more often than predicted (at alpha level 0.05). In these
  cases, the model underestimated the amount of overlap due to chance.}
  \item{P > 0.975}{The number of pairs for which overlapping occurred
  significantly less often than predicted (at alpha level 0.05). In these
  cases, the model overestimated the amount of overlap due to chance.}
}
}
\description{
Given a list of non-interacting pairs of individuals,
\code{song.AccuracyTest} calculates the difference between the actual amount
of chance overlap and the amount predicted by the model. This function can
be used to determine which of the randomization methods included in the
package (\code{\link{song.RandomizeSampleGaps}},
\code{\link{song.RandomizeKeepGaps}}, or
\code{\link{song.RandomizeKeepSongOrder}}) is best for a given system.
}
\details{
Non-interacting pairs (i.e. individuals that are separated by time or
space) should overlap some amount by chance. The accuracy of a song overlap
null model can be measured as the difference between this amount and the
prediction generated by the model.

When assembling a data set for this analysis, individuals should be
randomly assigned to pairs, and the start time of each individual's
performance should be randomly selected. These procedures ensure that any
overlap occuring during these "non-interactions" is due to chance alone.
It is important to note that the quality of the accuracy measurement depends
on the number of non-interacting pairs included in the analysis.
}
\seealso{
\code{\link{song.HowManyRand}} to evaluate the accuracy of a model as a
function of the number of randomizations.
}

