% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/song.io.R
\name{song.BuildIndiv}
\alias{song.BuildIndiv}
\title{Build a list of performance statistics for an individual.}
\usage{
song.BuildIndiv(ID, songs, start.record.time, end.record.time)
}
\arguments{
\item{ID}{A character string giving the name of the individual.}

\item{songs}{An n x 2 matrix or data frame containing the start and end
times of n songs. The first column contains the start times of each song;
the second column contains the end times of each song.}

\item{start.record.time}{A numeric value indicating the start time of the
performance.}

\item{end.record.time}{A numeric value indicating the end time of the
performance.}
}
\value{
\code{song.BuildIndiv} returns a list containing the following components:
\describe{
  \item{\code{ID}}{A character string giving the name of the individual.}
  \item{\code{start.record.time}}{A numeric value indicating the start time
  of the performance.}
  \item{\code{end.record.time}}{A numeric value indicating the end time of
  the performance.}
  \item{\code{songs}}{A matrix containing the start and end times of
  the songs.}
  \item{\code{songs.num}}{An integer value indicating the number of songs
  in the performance.}
  \item{\code{songs.length}}{A named numeric vector containing the durations
  of each song in the performance.}
  \item{\code{songs.total.length}}{A numeric value indicating the total time
  spent singing (i.e. \code{sum(songs.length)}).}
  \item{\code{gaps.num}}{An integer value indicating the number of
  inter-song intervals in the performance.}
  \item{\code{gaps.length}}{A named numeric vector containing the durations
  of each inter-song interval in the performance.}
  \item{\code{gaps.total.length}}{A numeric value indicating the total time
  spent silent (i.e. \code{sum(gaps.length)}).}
}
}
\description{
\code{song.BuildIndiv} builds a list of performance statistics for an
individual based on the start and end times of the individual's songs.
}
\examples{
songs <- matrix(c(0.0, 1.0,
                  3.0, 4.0,
                  7.0, 8.0,
                 10.0, 11.0), ncol=2, byrow=TRUE)
bird <- song.BuildIndiv(ID = "bird", songs, start.record.time = 0.0,
                        end.record.time = 12.0)
}
\seealso{
To extract performance statistics from interactions involving multiple
individuals, use \code{\link{song.FromDataObj}} (for data stored in a matrix
or data frame) or \code{\link{song.FromTextFile}} (for data stored in a
tab-delimited text file.
}

