% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/song.who.R
\name{song.CountWhichOverlap}
\alias{song.CountWhichOverlap}
\title{Count instances of overlap given two performances.}
\usage{
song.CountWhichOverlap(songlist, reference)
}
\arguments{
\item{songlist,reference}{n x 2 matrices or data frames containing the start
and end times of n songs. The first columns contain the start times of each
song; the second columns contain the end times of each song.}
}
\value{
\code{song.CountWhichOverlap} returns the total number of instances
in which songs in \code{songlist} lag behind and overlap songs in the
reference songlist, (\code{reference}).
}
\description{
\code{song.CountWhichOverlap} counts the number of instances in which songs
in one performance lag behind and overlap songs in another performance. This
function is an extension of \code{song.WhichSongWithSonglist}.
}
\examples{
reference <- matrix(c(0.0, 1.0,
                      3.0, 4.0,
                      7.0, 8.0,
                     10.0, 11.0), ncol=2, byrow=TRUE)
songlist <- matrix(c(0.5, 1.5,
                     4.0, 5.0,
                     6.5, 7.5,
                    10.5, 12.0), ncol=2, byrow=TRUE)
song.CountWhichOverlap(songlist, reference)
}
\seealso{
\code{\link{song.WhichSongsOverlap}} and
\code{\link{song.WhichSongWithSongList}} for determining the direction of
overlap
}

