% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/song.overlap.R
\name{song.DoSongsOverlap}
\alias{song.DoSongsOverlap}
\title{Determine whether two songs overlap}
\usage{
song.DoSongsOverlap(my.song1, my.song2)
}
\arguments{
\item{my.song1,mysong.2}{Numeric vectors containing the start and end times
 of songs 1 and 2 respectively}
}
\value{
The amount of time (in seconds) for which the two songs overlap
}
\description{
\code{song.DoSongsOverlap} determines whether two songs overlap, and returns
the duration of overlap in seconds
}
\examples{
s1 <- c(0.0, 1.0)
s2 <- c(0.5, 2.0)
print(song.DoSongsOverlap(s1, s2))

s1 <- c(0.0, 1.0)
s2 <- c(1.0, 2.0)
print(song.DoSongsOverlap(s1, s2))
}

