% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/song.dutycycle.R
\name{song.DutyCycleMethod}
\alias{song.DutyCycleMethod}
\title{The duty cycle method for calculating chance overlap.}
\usage{
song.DutyCycleMethod(indivs, dc.function = song.DutyCycleNum)
}
\arguments{
\item{indivs}{A list created by \code{\link{song.BuildAllIndivs}} or
\code{\link{song.ReadSongList}} that contains the performance statistics of
each individual.}

\item{dc.function}{The function to be used to calculate the expected amount
of overlap. \code{\link{song.DutyCycleNum}} returns the number of
overlapping songs; \code{\link{song.DutyCycleTime}} returns the duration of
overlap in seconds. If not specified, \code{song.DutyCycleNum} is used.}
}
\value{
\code{song.DutyCycleMethod} returns a data frame containing the
observed amount of overlap, the expected amount of overlap, the value of the
chi-squared test statistic, and the resulting p-value for each possible pair
of individuals.
}
\description{
\code{song.DutyCycleMethod} calculates the expected amount of chance overlap
for a given interaction using the duty cycle method (Ficken et al. 1974),
then performs a chi-squared test to compare the observed and expected
values.
}
\details{
The duty cycle method was first described by Ficken, Ficken, and Hailman
(1974) in their study of acoustic interference in two avian species:
red-eyed vireos and least flycatchers. Traditionally, this method returns
the expected number of overlapping songs, but can be modified to return the
total duration of overlap by specifying \code{\link{song.DutyCycleTime}} as
the \code{dc.function}.
}
\examples{
w <- song.BuildAllIndivs(wrens)
## Traditional method: number of overlapping songs
song.DutyCycleMethod(w)
## Modified method: duration of overlap
song.DutyCycleMethod(w, song.DutyCycleTime)
}
\references{
Ficken RW, Ficken MS, Hailman JP. 1974. Temporal pattern shifts to avoid
acoustic interference in singing birds. Science. 183:762-763.
}

