% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/song.io.R
\name{song.FromDataObj}
\alias{song.FromDataObj}
\title{Extract performance statistics from a matrix or data frame.}
\usage{
song.FromDataObj(songs, start.record.time = NA, end.record.time = NA)
}
\arguments{
\item{songs}{An n x 3 matrix or data frame containing the start time, end
time, and singer identity for each of n songs. The first column contains the
start times of each song; the second column contains the end times of each
song; the third column specifies the name of the individual that produced
the song.}

\item{start.record.time}{A numeric value indicating the start time of the
interaction. An optional parameter - if not specified, the minimum start
time is used (i.e. the start of the first song).}

\item{end.record.time}{A numeric value indicating the end time of the
interaction. An optional parameter - if not specified, the maximum end time
is used (i.e. the end of the last song).}
}
\value{
\code{song.FromDataObj} returns a list containing the following components
for each individual:
\describe{
  \item{\code{ID}}{A character string giving the name of the individual.}
  \item{\code{start.record.time}}{A numeric value indicating the start time
  of the interaction.}
  \item{\code{end.record.time}}{A numeric value indicating the end time of
  the interaction.}
  \item{\code{songs}}{A matrix containing the start and end times of
  each of the individual's songs.}
  \item{\code{songs.num}}{An integer value indicating the number of songs
  produced by the individual.}
  \item{\code{songs.length}}{A named numeric vector containing the durations
  of each song produced by the individual.}
  \item{\code{songs.total.length}}{A numeric value indicating the total time
  for which the individual was singing.}
  \item{\code{gaps.num}}{An integer value indicating the number of
  inter-song intervals in the individual's performance.}
  \item{\code{gaps.length}}{A named numeric vector containing the durations
  of each inter-song interval in the individual's performance.}
  \item{\code{gaps.total.length}}{A numeric value indicating the total time
  for which the individual was silent.}
}
}
\description{
\code{song.FromDataObj} builds a list of performance statistics from a
matrix or data frame containing the start time, end time, and singer
identity for each song in an interaction. This function is an extension of
\code{\link{song.BuildIndiv}}.
}
\examples{
m <- song.FromDataObj(manakins)
## specify the start and end of the recording
m <- song.FromDataObj(manakins, 100, 250)
}
\seealso{
\code{\link{song.FromTextFile}} for data stored in a tab-delimited text
file.
}

