% Generated by roxygen2 (4.0.1): do not edit by hand
\name{song.FromTextFile}
\alias{song.FromTextFile}
\title{Extract performance statistics from a text file.}
\usage{
song.FromTextFile(file, start.record.time = NA, end.record.time = NA)
}
\arguments{
\item{file}{The name of the file from which the data are to be read. If it
does not contain an absolute path, the file name is relative to the current
working directory.}

\item{start.record.time}{A numeric value indicating the start time of the
interaction. An optional parameter - if not specified, the minimum start
time is used (i.e. the start of the first song).}

\item{end.record.time}{A numeric value indicating the end time of the
interaction. An optional parameter - if not specified, the maximum end time
is used (i.e. the end of the last song).}
}
\value{
\code{song.FromTextFile} returns a list containing the following components
for each individual:
\describe{
  \item{\code{ID}}{A character string giving the name of the individual.}
  \item{\code{start.record.time}}{A numeric value indicating the start time
  of the interaction.}
  \item{\code{end.record.time}}{A numeric value indicating the end time of
  the interaction.}
  \item{\code{songs}}{A matrix containing the start and end times of
  each of the individual's songs.}
  \item{\code{songs.num}}{An integer value indicating the number of songs
  produced by the individual.}
  \item{\code{songs.length}}{A named numeric vector containing the durations
  of each song produced by the individual.}
  \item{\code{songs.total.length}}{A numeric value indicating the total time
  for which the individual was singing (i.e. \code{sum(songs.length)}).}
  \item{\code{gaps.num}}{An integer value indicating the number of
  inter-song intervals in the individual's performance.}
  \item{\code{gaps.length}}{A named numeric vector containing the durations
  of each inter-song interval in the individual's performance.}
  \item{\code{gaps.total.length}}{A numeric value indicating the total time
  for which the individual was silent (i.e. \code{sum(gaps.length)}).}
}
}
\description{
\code{song.FromTextFile} builds a list of performance statistics for each
individual from a tab-delimited text file containing the start time, end
time, and singer identity for each song in an interaction. This function is
an extension of \code{\link{song.BuildIndiv}}.
}
\details{
The input for this function is a tab-delimited text file (*.txt) containing
the start time, end time, and singer identity for each song during an
interaction. These data can be acquired using sound annotation software
packages such as \href{http://www.syrinxpc.com/}{Syrinx-PC} (J. Burt,
Seattle, WA, USA) or
\href{http://www.birds.cornell.edu/brp/raven/RavenOverview.html}{Raven}
(Cornell Laboratory of Ornithology, Ithaca, NY USA).
The following is an example input file:
\tabular{lll}{
start \tab end \tab ID \cr
0 \tab 5 \tab bird1 \cr
2 \tab 8 \tab bird2 \cr
10 \tab 15 \tab bird1 \cr
12 \tab 16 \tab bird2
}
}
\seealso{
\code{\link{song.FromDataObj}} for data stored in a matrix or
data frame.
}

