% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/song.overlap.R
\name{song.OneSongWithSongList}
\alias{song.OneSongWithSongList}
\title{Calculate the overlap of one song and a performance.}
\usage{
song.OneSongWithSongList(song, songlist)
}
\arguments{
\item{song}{A numeric vector containing the start and end times of a song.
The first value of the vector represents the start time of the song; the
second value represents the end time of the song.}

\item{songlist}{An n x 2 matrix or data frame containing the start and end
times of n songs. The first column contains the start times of each song;
the second column contains the end times of each song.}
}
\value{
\code{song.OneSongWithSongList} returns the total duration of overlap in
seconds.
}
\description{
\code{song.OneSongWithSongList} calculates the duration for which one
song overlaps the songs in a given performance. A song is considered
overlapping if it begins while another song is in progress. This function
is an extension of \code{\link{song.DoSongsOverlap}}.
}
\examples{
## overlapping song
song <- c(0.5, 2.0)
songlist <- matrix(c(0.0, 1.0,
                     3.0, 4.0,
                     7.0, 8.0,
                    10.0, 11.0), ncol=2, byrow=TRUE)
song.OneSongWithSongList(song, songlist)

## non-overlapping song
song <- c(1.0, 2.0)
songlist <- matrix(c(0.0, 1.0,
                     3.0, 4.0,
                     7.0, 8.0,
                    10.0, 11.0), ncol=2, byrow=TRUE)
song.OneSongWithSongList(song, songlist)
}

