% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/song.plotting.R
\name{song.PlotResultsDensity}
\alias{song.PlotResultsDensity}
\title{Plot the observed and expected amounts of song overlap.}
\usage{
song.PlotResultsDensity(results)
}
\arguments{
\item{results}{A list created by \code{\link{song.Simulate}} that contains
the observed and expected amounts of overlap for an interaction.}
}
\value{
\code{song.PlotResultDensity} returns a \code{\link{ggplot}} object in which
each panel of the plot represents a pairwise interaction. Within each panel,
the observed amount of overlap (dashed line) is plotted relative to the null
distribution (filled curve). The p-value corresponding to each pairwise
interaction appears in the upper right corner of each panel.
}
\description{
\code{song.PlotResultsDensity} creates a \code{\link{ggplot}} object that
depicts the observed amount of overlap relative to the expected amount for
each possible pair of individuals. This function can be used to visualize
the output generated by \code{\link{song.Simulate}}.
}
\examples{
## Black-capped chickadees
c <- song.BuildAllIndivs(chickadees)
c.rand <- song.Simulate(c, 100, song.TimeOverlap,
                        song.RandomizeSampleGaps)
song.PlotResultsDensity(c.rand)

## Rufous-and-white wrens
w <- song.BuildAllIndivs(wrens)
w.rand <- song.Simulate(w, 100, song.TimeOverlap,
                        song.RandomizeKeepSongOrder)
song.PlotResultsDensity(w.rand)

## Long-tailed manakins
m <- song.BuildAllIndivs(manakins)
m.rand <- song.Simulate(m, 100, song.TimeOverlap,
                        song.RandomizeSampleGaps)
song.PlotResultsDensity(m.rand)
}

