% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/song.plotting.R
\name{song.PlotSongs}
\alias{song.PlotSongs}
\title{Plot the relative timing of songs during an interaction.}
\usage{
song.PlotSongs(indivs, start.time = NA, end.time = NA)
}
\arguments{
\item{indivs}{A list created by \code{\link{song.BuildAllIndivs}} or
\code{\link{song.ReadSongList}} that contains the performance statistics of
each individual.}

\item{start.time}{A numeric value indicating the start time of the
interaction in seconds. An optional parameter - if not specified, the
minimum start time is used (i.e. the start of the first song).}

\item{end.time}{A numeric value indicating the end time of the
interaction in seconds. An optional parameter - if not specified, the
maximum end time is used (i.e. the end of the last song).}
}
\description{
\code{song.PlotSongs} creates a \code{\link{ggplot}} object depicting the
relative timing of songs during an interaction. Each song is represented by
a line segment plotted along a time axis. Songs are color-coded and
plotted by individual.
}
\details{
The optional parameters \code{start.time} and \code{end.time} set the limits
of the x-axis of the plot. If specified, these parameters can be used to
view select portions of an interaction. This practice often results in
the following warning message: "Removed ## rows containing missing values
(geom_segment)." This warning can be safely ignored.
}
\examples{
c <- song.BuildAllIndivs(chickadees)

## Plot the entire interaction
song.PlotSongs(c)

## Plot the songs occuring between 02m 30s and 02m 50s
song.PlotSongs(c, 150, 170)
}

