% Generated by roxygen2 (4.0.1): do not edit by hand
\name{song.RandomizeKeepGaps}
\alias{song.RandomizeKeepGaps}
\title{Generate a randomized performance using the KeepGaps method.}
\usage{
song.RandomizeKeepGaps(indiv)
}
\arguments{
\item{indiv}{A list created using \code{\link{song.BuildAllIndivs}} or
\code{\link{song.ReadSongList}} that contains the performance statistics of
the individual.}
}
\value{
\code{song.RandomizeKeepGaps} returns an n x 2 matrix containing the start
and end times of the songs in the new randomized performance.
}
\description{
\code{song.RandomizeKeepGaps} generates a randomized performance based on
an individual's observed singing behavior. The KeepGaps method preserves the
individual's natural song and inter-song interval durations, while
randomizing the order of the songs and intervals. This function can be used
with \code{\link{song.Simulate}} to calculate the expected amount of chance
overlap in an interaction.
}
\examples{
m <- song.BuildAllIndivs(manakins)
song.RandomizeKeepGaps(m$pairA)
}
\seealso{
Other randomization.functions: \code{\link{song.RandomizeKeepSongOrder}};
  \code{\link{song.RandomizeSampleGaps}}
}

