% Generated by roxygen2 (4.0.1): do not edit by hand
\name{song.RandomizeSampleGaps}
\alias{song.RandomizeSampleGaps}
\title{Generate a randomized performance using the SampleGaps method.}
\usage{
song.RandomizeSampleGaps(indiv)
}
\arguments{
\item{indiv}{A list created using \code{\link{song.BuildAllIndivs}} or
\code{\link{song.ReadSongList}} that contains the performance statistics of
the individual.}
}
\value{
\code{song.RandomizeSampleGaps} returns an n x 2 matrix containing the start
and end times of the songs in the new randomized performance.
}
\description{
\code{song.RandomizeSampleGaps} generates a randomized performance based on
an individual's observed singing behavior. The SampleGaps method preserves
the individual's natural song durations, while randomizing both the order of
the songs and the duration of the inter-song intervals. This function can be
used with \code{\link{song.Simulate}} to calculate the expected amount of
chance overlap in an interaction.
}
\examples{
c <- song.BuildAllIndivs(chickadees)
song.RandomizeSampleGaps(c$bird)
}
\seealso{
Other randomization.functions: \code{\link{song.RandomizeKeepGaps}};
  \code{\link{song.RandomizeKeepSongOrder}}
}

