% Generated by roxygen2 (4.0.1): do not edit by hand
\name{song.ReadRecitals}
\alias{song.ReadRecitals}
\title{Reads a file containing the recitals for the birds}
\usage{
song.ReadRecitals(file.recitals, start.record.time = NA,
  end.record.time = NA)
}
\arguments{
\item{file.recitals}{A file containing the start time and end time of all the songs}

\item{start.record.time}{Optional parameter to set a particular start time for the recording}

\item{end.record.time}{Optional parameter to set a particular end time for the recording}
}
\value{
A list of birds along with their recital
}
\description{
Reads a file containing the recitals for the birds
}
\examples{
\dontrun{
Rec <- song.ReadRecitals("test.txt", 0.1, 100)
}
}

