% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/song.simulate.R
\name{song.Simulate}
\alias{song.Simulate}
\title{Compare observed overlap to chance expectations.}
\usage{
song.Simulate(indivs, num.rand = 100, overlap.function = "song.TimeOverlap",
  randomize.function = "song.RandomizeSampleGaps")
}
\arguments{
\item{indivs}{A list created by \code{\link{song.BuildAllIndivs}} or
\code{\link{song.ReadSongList}} that contains the performance statistics of
each individual.}

\item{num.rand}{A numeric value indicating the desired number of
randomizations.}

\item{overlap.function}{The function to be used to calculate the amount of
overlap. \code{\link{song.TimeOverlap}} returns the duration of overlap
in seconds; \code{\link{song.NumOverlap}} returns the number of overlapping
songs.}

\item{randomize.function}{The function to be used to generate the randomized
performances for each individual. Options include
\code{\link{song.RandomizeSampleGaps}},
\code{\link{song.RandomizeKeepGaps}}, and
\code{\link{song.RandomizeKeepSongOrder}}.}
}
\value{
\code{song.Simulate} returns a list containing the following
components:
\describe{
  \item{observed}{A matrix containing the amount of overlap in the observed
  interaction for each possible pair of individuals.}
  \item{expected}{A matrix containing the average amount of overlap in the
  randomized interactions for each possible pair of individuals. These
  values represent the expected amounts of overlap due to chance.}
  \item{p.values}{A matrix containing the p-values associated with each
  possible pair of individuals.}
  \item{randomized}{A three-dimensional array containing the amount of
  overlap occurring in each randomized interaction. Each "slice" is a
  matrix containing the amount of overlap for each possible pair of
  individuals. These values make up the null distribution.}
  \item{overlap.method}{The function used to calculate the amount of
  overlap.}
  \item{randomize.method}{The function used to generate the randomized
  interactions.}
}
}
\description{
\code{song.Simulate} calculates the expected amount of overlap due to chance
for a given interaction, then compares the observed amount of overlap to
this expectation.
}
\details{
\code{song.Simulate} generates a user-defined number of randomized
interactions for each pair of individuals. The function then calculates the
amount of overlap occuring in each of these randomized interactions. These
values represent the null distribution (i.e. the amount of overlap expected
due to chance). For each pair of individuals, the p-value is calculated as
the percentage of randomized interactions in which the amount of overlap is
greater than the amount of overlap in the observed interaction.
}
\examples{
c <- song.BuildAllIndivs(chickadees)
c.rand <- song.Simulate(c, num.rand = 100, song.TimeOverlap,
                        song.RandomizeSampleGaps)
## Duration of overlap in the observed interaction
c.rand$observed
## Duration of overlap expected due to chance
c.rand$expected
## How does observed overlap compare to chance?
c.rand$p.values
}
\seealso{
\code{\link{song.PlotResultsDensity}} to visualize the output of
\code{song.Simulate}.
}

