% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/song.who.R
\name{song.WhichSongsOverlap}
\alias{song.WhichSongsOverlap}
\title{Determine the direction of song overlap.}
\usage{
song.WhichSongsOverlap(song1, song2)
}
\arguments{
\item{song1,song2}{Numeric vectors containing the start and end times
of songs 1 and 2 respectively. The first value in each vector represents the
start time of the song; the second value represents the end time of the
song.}
}
\value{
If \code{song2} begins during \code{song1} (i.e.\code{song1} is leading),
the function returns a value of 1. If \code{song1} begins during
\code{song2} (i.e. \code{song2} is leading), the function returns a value of
-1. If the songs are non-overlapping, the function returns a value of 0.
}
\description{
\code{song.WhichSongsOverlap} determines whether two songs overlap and
assigns a value (1 or -1) to indicate the order of the overlapping songs.
}
\examples{
## songs overlap and song1 is leading
song1 <- c(0.0, 1.0)
song2 <- c(0.5, 2.0)
song.WhichSongsOverlap(song1,song2)

## songs overlap and song1 is lagging
song1 <- c(1.0, 3.0)
song2 <- c(0.0, 2.0)
song.WhichSongsOverlap(song1,song2)

## songs do not overlap
song1 <- c(0.0, 1.0)
song2 <- c(2.0, 3.0)
song.WhichSongsOverlap(song1,song2)
}
\seealso{
\code{\link{song.CountWhichOverlap}} for counting instances of
overlap given two performances, and \code{\link{song.WhoOverlaps}} for
counting instances of overlap by individual.
}

