% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/song.who.R
\name{song.WhichSongWithSongList}
\alias{song.WhichSongWithSongList}
\title{Determine the direction of overlap for one song and a performance.}
\usage{
song.WhichSongWithSongList(song, songlist)
}
\arguments{
\item{song}{A numeric vector containing the start and end times of a song.
The first value of the vector represents the start time of the song; the
second value represents the end time of the song.}

\item{songlist}{An n x 2 matrix or data frame containing the start and end
times of n songs. The first column contains the start times of each song;
the second column contains the end times of each song.}
}
\value{
If \code{song} begins during any song in \code{songlist}, the function
returns a value of 1. Otherwise (i.e. if there is no overlap, or overlapping
occurs in the opposite direction), the function returns a value of 0.
}
\description{
\code{song.WhichSongWithSongList} determines whether a song overlaps and
lags behind any song in a given performance. This function is an extension
of \code{\link{song.WhichSongsOverlap}}.
}
\examples{
## song overlaps and lags behind a song in songlist
song <- c(0.5, 2.0)
songlist <- matrix(c(0.0, 1.0,
                     3.0, 4.0,
                     7.0, 8.0,
                    10.0, 11.0), ncol=2, byrow=TRUE)
song.WhichSongWithSongList(song, songlist)

## song overlaps but starts before a song in songlist
song <- c(2.5, 4.0)
songlist <- matrix(c(0.0, 1.0,
                     3.0, 4.0,
                     7.0, 8.0,
                    10.0, 11.0), ncol=2, byrow=TRUE)
song.WhichSongWithSongList(song, songlist)
}

