% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{women}
\alias{women}
\title{Average Heights and Weights for American Women}
\format{The women data frame contains 15 observations on 2 variables.
\describe{
  \item{height}{Height, in inches}
  \item{weight}{Weight in lbs}
}}
\source{
\url{http://www.worldalmanac.com/}
}
\usage{
{women}
}
\description{
This data provides a small sample of average heights and weights
for American women aged 30–39 and some random data from the normal
distribution.
}
\keyword{datasets}
