% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/song.datasets.R
\name{wrens}
\alias{wrens}
\title{A sample interaction from rufous-and-white wrens.}
\format{A data frame with 50 rows and 3 variables}
\source{
Mennill D, Burt J, Fristrup K, Vehrencamp S. 2006. Accuracy of an acoustic
location system for monitoring the position of duetting songbirds in
tropical forest. J Acoust Soc Am. 119:2832-2839.

Mennill DJ, Vehrencamp SL. 2008. Context-dependent functions of avian duets
revealed by microphone-array recordings and multispeaker playback. Curr
Biol. 18:1314-1319.
}
\description{
A data set containing the start and end times of each song produced during
an interaction between a duetting pair of rufous-and-white wrens. These data
were collected using an 8-channel microphone array designed to passively
sample the natural duetting behavior of breeding partners as they move
around their territories (Mennill et al. 2006; Mennill & Vehrencamp 2008).
}

