% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/song.R
\docType{data}
\name{chickadees}
\alias{chickadees}
\title{An interaction between a black-capped chickadee and a playback stimulus.}
\format{A data frame with 50 rows and 3 variables}
\source{
Baker TM, Wilson DR, Mennill DJ. 2012. Vocal signals predict attack
during aggressive interactions in black-capped chickadees. Anim Behav. 84:
965-974.
}
\usage{
chickadees
}
\description{
A data set containing the start and end times of each song produced during
an interaction between a black-capped chickadee and a fixed-interval
playback stimulus. These data are from a study by Baker, Wilson and Mennill
(2012) designed to evaluate whether song overlap predicts the attack of a
rival in this species.
}
\keyword{datasets}

