% Generated by roxygen2 (4.0.1): do not edit by hand
\name{song.DoSongsOverlap}
\alias{song.DoSongsOverlap}
\title{Checks whether two songs are overlapping, and returns time they overlap}
\usage{
song.DoSongsOverlap(my.song1, my.song2)
}
\arguments{
\item{my.song1}{A vector containing starting (my.song1[1]) and stopping time (my.song1[2])}

\item{my.song2}{A vector containing starting (my.song2[1]) and stopping time (my.song2[2])}
}
\value{
The amount of time the two songs overlap
}
\description{
Checks whether two songs are overlapping, and returns time they overlap
}
\examples{
s1 <- c(0.0, 1.0)
s2 <- c(0.5, 1.25)
print(song.DoSongsOverlap(s1, s2))

s1 <- c(0.0, 0.5)
s2 <- c(0.5, 1.25)
print(song.DoSongsOverlap(s1, s2))
}

