% Generated by roxygen2 (4.0.1): do not edit by hand
\name{song.DutyCycleNum}
\alias{song.DutyCycleNum}
\title{Predict the number of overlapping songs using the duty cycle method.}
\usage{
song.DutyCycleNum(target, reference)
}
\arguments{
\item{reference,target}{Lists created using \code{\link{song.FromDataObj}}
or \code{\link{song.FromTextFile}} that contain the performance statistics
of two individuals. The amount of overlap is calculated for the target
individual with respect to the reference individual.}
}
\value{
\code{song.DutyCycleNum} returns a list with class "\code{htest}"
containing the following components:
\describe{
  \item{\code{statistic}}{the value of the chi-squared test statistic.}
  \item{\code{parameter}}{the degrees of freedom of the approximate
  chi-squared distribution of the test statistic.}
  \item{\code{p.value}}{the p-value for the test.}
  \item{\code{method}}{a character string indicating the type of test
  performed.}
  \item{\code{data.name}}{a character string giving the names of the
  individuals included in the analysis.}
  \item{\code{observed}}{a numeric vector containing the observed number of
  overlapping and non-overlapping songs.}
  \item{\code{expected}}{a numeric vector containing the expected number of
  overlapping and non-overlapping songs.}
  \item{\code{residuals}}{the Pearson residuals, (\code{observed} -
  \code{expected}) / sqrt(\code{expected}).}
}
}
\description{
\code{song.DutyCycleNum} calculates the expected number of overlapping and
non-overlapping songs, then performs a chi-squared test to compare the
observed and expected values. This function is to be used with
\code{\link{song.DutyCycleMethod}}.
}
\details{
The duty cycle method was first described by Ficken, Ficken, and Hailman
(1974) in their study of acoustic interference in two avian species:
red-eyed vireos and least flycatchers. According to this method, the
expected number of flycatcher songs that overlap vireo songs is equal to the
total number of flycatcher songs multiplied by the probability that the
vireo is singing.
}
\examples{
c <- song.FromDataObj(chickadees)
song.DutyCycleNum(c$bird, c$playback)
}
\references{
Ficken RW, Ficken MS, Hailman JP. 1974. Temporal pattern shifts to avoid
acoustic interference in singing birds. Science. 183:762-763.
}
\seealso{
\code{\link{song.DutyCycleTime}} for calculating the \emph{duration} of
overlap rather than the number of overlapping songs.
}

