% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/song.overlap.R
\name{song.NumOverlap}
\alias{song.NumOverlap}
\title{Count the total number of songs that overlap.}
\usage{
song.NumOverlap(target, reference)
}
\arguments{
\item{target,reference}{n x 2 matrices or data frames containing the start
and end times of n songs. The first columns contain the start times of each
song; the second columns contain the end times of each song. The amount of
overlap is calculated for the \code{target} with respect to the
\code{reference}.}
}
\value{
\code{song.NumOverlap} returns the total number of songs that
overlap.
}
\description{
\code{song.NumOverlap} counts the total number of songs in one performance
that overlap and lag behind the songs in another. This function is an extension of
\code{\link{song.OneSongWithSongList}}.
}
\examples{
reference <- matrix(c(0.0, 1.0,
                      3.0, 4.0,
                      7.0, 8.0,
                     10.0, 11.0), ncol=2, byrow=TRUE)
target <- matrix(c(0.2, 1.5,
                     4.0, 5.0,
                     6.5, 7.5,
                    10.5, 12.0), ncol=2, byrow=TRUE)
song.NumOverlap(target, reference)
}
\seealso{
\code{\link{song.DoSongsOverlap}} for comparing two songs,
\code{\link{song.OneSongWithSongList}} for comparing one song to a
performance, and \code{\link{song.TimeOverlap}} for calculating the
\emph{duration} of overlap rather than the number of overlapping songs.
}

