% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/song.who.R
\name{song.WhoOverlaps}
\alias{song.WhoOverlaps}
\title{Count the instances of overlap by individual.}
\usage{
song.WhoOverlaps(indivs)
}
\arguments{
\item{indivs}{A list created using \code{\link{song.BuildAllIndivs}} or
\code{\link{song.ReadSongList}} that contains the performance statistics of
each individual.}
}
\value{
\code{song.WhoOverlaps} returns a data frame containing the number
of instances of overlap for each pairwise interaction, specifying the
"overlappers" and reference individuals in each case.
}
\description{
\code{song.WhoOverlaps} counts the number of instances in which each
individual overlaps each other individual's songs. This function is an
extension of \code{\link{song.CountWhichOverlap}}.
}
\examples{
c <- song.BuildAllIndivs(chickadees)
song.WhoOverlaps(c)

w <- song.BuildAllIndivs(wrens)
song.WhoOverlaps(w)

m <- song.BuildAllIndivs(manakins)
song.WhoOverlaps(m)
}
\seealso{
\code{\link{song.CountWhichOverlap}}
}

