% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/song.who.R
\name{song.WhoOverlaps}
\alias{song.WhoOverlaps}
\title{Count the instances of overlap by individual.}
\usage{
song.WhoOverlaps(birds)
}
\arguments{
\item{birds}{A list created using \code{song.BuildSongList} or
\code{song.ReadSongList} that contains the performances of each individual.}
}
\value{
\code{song.WhoOverlaps} returns a data frame containing the number
of instances of overlap for each pairwise interaction, specifying the
"overlappers" and reference individuals in each case.
}
\description{
\code{song.WhoOverlaps} counts the number of instances in which each
individual overlaps each other individual's songs. This function is an
extension of \code{song.CountWhichOverlap}.
}
\examples{
w <- song.BuildSongList(wrens)
song.WhoOverlaps(w)
}
\seealso{
\code{\link{song.CountWhichOverlap}}
}

